+++ src/ace_time/Epoch.cpp
@@ -5,6 +5,6 @@ namespace ace_time {
 int16_t Epoch::sCurrentEpochYear = 2050;
 
 // Number of days from 2000-01-01 to 2050-01-01: 50*365 + 13 leap days = 18263.
-int32_t Epoch::sDaysToCurrentEpochFromConverterEpoch = 18263;
+int32_t Epoch::sDaysToCurrentEpochFromInternalEpoch = 18263;
 
 }